Revert "feat: ft to expand tkn_list; new file bc ft limit"
This reverts commit 8d1cb732d6faba973ef6a3d1e608d967e555f244.
